Make gamepad deadzone variables shorts

+++ src/c_cmds.c
@@ -2703,12 +2703,12 @@ static void gp_deadzone_cvars_func2(char *cmd, char *parm1, char *parm2, char *p
         if (M_StringCompare(cmd, stringize(gp_deadzone_left)))
         {
             gp_deadzone_left = BETWEENF(gp_deadzone_left_min, value, gp_deadzone_left_max);
-            gamepadleftdeadzone = (int)(gp_deadzone_left * (float)SHRT_MAX / 100.0f);
+            gamepadleftdeadzone = (short)(gp_deadzone_left * (float)SHRT_MAX / 100.0f);
         }
         else
         {
             gp_deadzone_right = BETWEENF(gp_deadzone_right_min, value, gp_deadzone_right_max);
-            gamepadrightdeadzone = (int)(gp_deadzone_right * (float)SHRT_MAX / 100.0f);
+            gamepadrightdeadzone = (short)(gp_deadzone_right * (float)SHRT_MAX / 100.0f);
         }
         M_SaveCVARs();
     }
